+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # [CHANGELOG](https://keepachangelog.com/en/1.0.0/)
 
-## [5.13.1](https://github.com/gdsfactory/gdsfactory/pull/552)
+## [5.14.0](https://github.com/gdsfactory/gdsfactory/pull/553)
 
 - Port has optional width that can get it from the cross_section
+- Deprecate Port.midpoint. Use Port.center instead.
+- Deprecate Port.position. Use Port.center instead.
+- Deprecate Port.angle. Use Port.orientation instead.
 
 ## [5.13.0](https://github.com/gdsfactory/gdsfactory/pull/552)
 
@@ -616,7 +619,7 @@
   - add shear_angle attribute to Port
   - Update test_shear_face_path.py
 
-- remove default port width, layer and midpoint
+- remove default port width, layer and center
 
 ## [5.0.7](https://github.com/gdsfactory/gdsfactory/pull/281)